refactor: 컨트롤러 이전 view 레이어서는 worker 내부의 Request DTO를 몰라도 실행가능하도록 변경
